Added logobj functionality, updated readme and sample
* Can now define the format for what objects are returned from logobj.
This pairs members with format strings.
* included a special case for author and committer dates to be returned
as Date() objects. The special casing is handled only at the top of the
file, should be easy enough to extend.
* removed -1 param from the _command methods slice of stdout - it was
removing the last character from the output which broke formatting when
the last param was an empty string.
* added a format param to log to allow users to specify the layout of
the returned array - I made the default value the format it had before
I started modifying the codebase.
* reorganized the file to encourage hacking - the top is easily
modifiable, followed by what’s exposed and then finally internal
functionality that may be less tempting to mess with.